+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.0.2
+
+- New setting to hide functions nested under methods or other functions
+- Display parent hierarchy in the symbol description right to the symbol name
+
 ## 1.0.1
 
 - Initial release